requirements.txt
Flask==2.0.2
gunicorn==20.1.0
PyYAML==5.3.1
xmltodict==0.12.0
requests==2.26.0
SQLAlchemy==1.4.27
pyinstaller==4.5.1
flask-swagger-ui==4.11.1